refactor: move react specific utils back to reactjs
